fix(web): finish mirror readiness prerequisite wiring
Summary:
- extract the workflow repository prerequisite mapper into its own web
  module so the workflow data layer stays within file-budget limits
- update workflow page components to import the shared prerequisite type
  from the new module while keeping the same backend-backed behavior

Rationale:
- the readiness fix needs to satisfy the repo's frontend structure gate,
  not only the runtime behavior and type checks
- isolating the prerequisite contract makes the UI wiring easier to keep
  aligned with the backend workflow prerequisite endpoint

Tests:
- PATH=/home/yuzhong/.nvm/versions/node/v22.22.1/bin:$PATH cd web && pnpm exec vitest run src/lib/features/workflows/data.test.ts
- PATH=/home/yuzhong/.nvm/versions/node/v22.22.1/bin:$PATH cd web && pnpm exec svelte-check --fail-on-warnings
- PATH=/home/yuzhong/.nvm/versions/node/v22.22.1/bin:$PATH corepack pnpm --dir web run ci
